+++ src/common/allowedOrigins.ts
@@ -1,0 +1,7 @@
+const ALLOWED_ORIGINS = [
+  'http://localhost:8080',
+  'http://127.0.0.1:5500',
+  'https://admin.socket.io',
+];
+
+export default ALLOWED_ORIGINS;
