+++ src/handlers/game/index.ts
@@ -1,0 +1,13 @@
+import handleConnection from './handleConnection';
+import handleStartHandshake from './handleStartHandshake';
+import handleFinishHandshake from './handleFinishHandshake';
+import handleStartGameStatusUpdate from './handleStartGameStatusUpdate';
+import handleFinishGameStatusUpdate from './handleFinishGameStatusUpdate';
+
+export {
+  handleConnection,
+  handleStartHandshake,
+  handleFinishHandshake,
+  handleStartGameStatusUpdate,
+  handleFinishGameStatusUpdate,
+};
