+++ src/handlers/user/handleJoinRoom.ts
@@ -1,57 +1,14 @@
-/* eslint-disable no-param-reassign */
-import { log, MAX_PLAYERS, randomID } from '../common/constants';
-import { MatchEvent } from '../common/enums';
-import {
-  ExtendedError,
-  IO,
-  MiClientSocket,
-  MiServerSocket,
-  Player,
-} from '../common/types';
-import InMemorySessionStore from '../utils/InMemorySessionStore';
+import { IO, MiClientSocket, MiServerSocket } from '../../common/interfaces';
+import InMemorySessionStore from '../../utils/InMemorySessionStore';
+import { MAX_PLAYERS } from '../../common/constants';
+import { MatchEvent } from '../../common/enums';
 
-export const beforeConnectionOrReconnection = (
-  socket: MiClientSocket,
-  next: (err?: ExtendedError | undefined) => void,
-  sessionStore: InMemorySessionStore,
-  // eslint-disable-next-line consistent-return
-) => {
-  const { sessionID } = socket.handshake.auth;
-  if (sessionID) {
-    const session = sessionStore.findSession(sessionID);
-    if (session) {
-      socket.sessionID = sessionID;
-      socket.userID = session.userID;
-      socket.nickname = session.nickname;
-      socket.roomCode = session.roomCode;
-      socket.gameStatus = session.gameStatus;
-      log('si session found, sessionID:', sessionID);
-      return next();
-    }
-  }
-  const { nickname } = socket.handshake.auth;
-  if (!nickname) {
-    return next(new Error('invalid nickname'));
-  }
-  const roomCode = socket.handshake.query.gameCode;
-  if (!roomCode) {
-    return next(new Error('invalid gamecode'));
-  }
-  log('no session found');
-  socket.nickname = nickname;
-  socket.roomCode = roomCode as string;
-  socket.sessionID = randomID();
-  socket.userID = randomID();
-  socket.gameStatus = undefined;
-  next();
-};
-
-export const createOrJoinRoom = async (
+const handleJoinRoom = async (
   io: IO,
   socket: MiClientSocket,
   sessionStore: InMemorySessionStore,
 ) => {
-  const usersInRoom: Player[] = [];
+  const usersInRoom = [];
   const socketsAlreadyInRoom: MiServerSocket[] = (await ((await io
     .in(socket.roomCode)
     .fetchSockets()) as unknown)) as MiServerSocket[];
@@ -138,19 +95,4 @@ export const createOrJoinRoom = async (
   }
 };
 
-export const handleUserDisconnection = (
-  socket: MiClientSocket,
-  sessionStore: InMemorySessionStore,
-) => {
-  socket.in(socket.roomCode).emit('player disconnected', {
-    userID: socket.userID,
-    nickname: socket.nickname,
-  });
-  sessionStore.saveSession(socket.sessionID!, {
-    userID: socket.userID!,
-    nickname: socket.nickname,
-    roomCode: socket.roomCode,
-    connected: false,
-    gameStatus: socket.gameStatus,
-  });
-};
+export default handleJoinRoom;
