+++ src/handlers/user/index.ts
@@ -1,0 +1,4 @@
+import handleJoinRoom from './handleJoinRoom';
+import handleDisconnection from './handleDisconnection';
+
+export { handleJoinRoom, handleDisconnection };
